aws-wiki-backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-storages==1.14.2
djangorestframework==3.15.1
idna==3.7
Markdown==3.6
mysqlclient==2.2.4
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
cx_Oracle
mysql-connector-python